Convert from macros to Rules and Service Providers and load Helpers
